% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CArecs.R, R/allDATRAS_1965-2014.R
\docType{data}
\name{CArecs}
\alias{CArecs}
\title{Age-based species information for ICES NS-IBTS survey from 1965-2014}
\format{A data frame with 321965 rows and 24 variables:
\describe{
  \item{Survey}{price, in US dollars}
  \item{Quarter}{weight of the diamond, in carats}
  \item{Ship}{price, in US dollars}
  \item{Gear}{weight of the diamond, in carats}
  \item{SweepLngt}{price, in US dollars}
  \item{GearExp}{weight of the diamond, in carats}
  \item{DoorType}{price, in US dollars}
  \item{StNo}{weight of the diamond, in carats}
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
}}
\source{
\url{https://datras.ices.dk/WebServices/Webservices.aspx}

\url{https://datras.ices.dk/WebServices/Webservices.aspx}
}
\usage{
CArecs

CArecs
}
\description{
A dataset containing age-based species information for ICES North Sea
international bottom trawl survey.

A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

