% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flr2xml.R
\name{flr2xml}
\alias{flr2xml}
\title{flr2xml.R
Translate FLCore "FLRobject" into xml template ready for ICES Stock Assessment Graphs}
\usage{
flr2xml(FLRobject, AssessmentYear, icesID, Fage, RecruitmentAge,
  catchType = c("official", "ICES estimates", "Model estimates"),
  outputDirectory = "~/", FLim = NULL, Fpa = NULL, Bpa = NULL,
  Blim = NULL, FMSY = NULL, MSYBtrigger = NULL, Fmanagement = NULL,
  Bmanagement = NULL, RecruitmentLength = NULL, Custom = FALSE,
  CustomName = FALSE, ...)
}
\arguments{
\item{AssessmentYear}{icesID
Fage
RecruitmentAge
outputDirectory
FLim = NULL
Fpa = NULL
Bpa = NULL
Blim = NULL
FMSY = NULL
MSYBtrigger = NULL
Fmanagement = NULL
Bmanagement  = NULL
RecruitmentLength = NULL
Custom = FALSE
CustomName = FALSE}
}
\value{
none
}
\description{
flr2xml.R
Translate FLCore "FLRobject" into xml template ready for ICES Stock Assessment Graphs
}
\details{
none
}
\examples{
\dontrun{
 needList <- c("XML", "RCurl", "data.table", "FLCore")
 new.packages <- needList[!(needList \%in\% installed.packages()[,"Package"])]
 if("FLCore" \%in\% new.packages) {
  install.packages("FLCore", repos="http://flr-project.org/R")
  new.packages <- new.packages[new.packages != "FLCore"]
 }
 if(length(new.packages)) install.packages(new.packages)
 lapply(needList, require, c = T)

 urls <- ("https://community.ices.dk/ExpertGroups/WGMIXFISH-NS/2015\%20Meeting\%20docs/06.\%20Data/FLStock\%20objects/North\%20Sea/standardised\%20RData/PLE-NS.RData")
 tmpFile <- tempfile()
 download.file(urls, destfile = tmpFile, mode = "wb")
 load(tmpFile)
 tt <- load(tmpFile)
 assign(x = "FLRobject", value = get(tt))

 flr2xml(FLRobject = FLRobject, AssessmentYear = 2015,
   icesID = "ple-nsea", Fage = c(2:6),
   Fmanagement = 0.3 , Bmanagement = 230000,
   RecruitmentAge = 1, catchType = "ICES Estimates",
   outputDirectory = "~/Data Products/FLR2XML/")

}
}
\seealso{
none
}
\keyword{none}

