% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linkRectangle.R
\name{getAreaName}
\alias{getAreaName}
\title{linkRect.R
Link ICES rectangles with ICES areas}
\usage{
getAreaName(areaList)
}
\arguments{
\item{surveyName,}{startYear, endYear, startQuarter, endQuarter, path = "~/", inParallel}
}
\value{
none
}
\description{
linkRect.R
Link ICES rectangles with ICES areas
}
\details{
the update is slow, avoiding straining the server or client.
  please allow this call to run overnight for a complete upgrade.
}
\examples{
\dontrun{
 getDATRAS(surveyName = "NS-IBTS", startYear = 2005, endYear = 2006, startQuarter = 1,
 endQuarter = 4, path = ".", inParallel = TRUE)
}
}
\seealso{
\code{\link{cacheDATRAS}}, a\code{\link{loadDATRAS}}, \code{\link{getCAfun}},
\code{\link{getHLfun}}, and \code{\link{getHHfun}}
}
\keyword{DATRAS,}
\keyword{download,}
\keyword{survey\}

