% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getCAfun.R
\name{getCAfun}
\alias{getCAfun}
\title{Datras Function}
\usage{
getCAfun(survey, startyear, endyear, startquarter, endquarter,
  parallel = FALSE)
}
\arguments{
\item{x}{}
}
\value{
none
}
\description{
More descriptive name
}

