% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDATRAS.R
\name{getDATRAS}
\alias{getDATRAS}
\title{getDATRASfun.R
Extracts age data files from DATRAS}
\usage{
getDATRAS(record, survey, startyear, endyear, quarters, parallel = FALSE,
  cores = NULL)
}
\arguments{
\item{record,}{survey, startyear, endyear, startquarter, endquarter, parallel = FALSE, cores = NULL}
}
\value{
none
}
\description{
getDATRASfun.R
Extracts age data files from DATRAS
}
\details{
the update is slow, avoiding straining the server or client.
  please allow this call to run overnight for a complete upgrade.
}
\examples{
\dontrun{
 getDATRAS(record = "HH",
           survey="NS-IBTS",
           startyear = 2010,
           endyear = 2011,
           quarters = 1,
           parallel = TRUE,
           cores = 4)
}
}
\seealso{
none
}
\keyword{DATRAS,}
\keyword{age,}
\keyword{download,}
\keyword{length}
\keyword{survey,}

