% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datrasFunctions.R, R/getHHfun.R
\name{getHHfun}
\alias{getHHfun}
\title{Datras Function}
\usage{
getHHfun(survey, startyear, endyear, startquarter, endquarter,
  parallel = FALSE)

getHHfun(survey, startyear, endyear, startquarter, endquarter,
  parallel = FALSE)
}
\arguments{
\item{survey,}{startyear, endyear, startquarter, endquarter, parallel}

\item{x}{}
}
\value{
none

none
}
\description{
More descriptive name

More descriptive name
}

