% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryTable.R
\name{getSummaryTable}
\alias{getSummaryTable}
\title{getSummaryTable.R
ICES standard graph data extraction}
\usage{
getSummaryTable(year = 2015)
}
\arguments{
\item{year}{The year of summary data to be extracted}
}
\value{
none
}
\description{
getSummaryTable.R
ICES standard graph data extraction
}
\details{
none
}
\examples{
\dontrun{
d <- getSummaryTable()
}
}
\seealso{
none
}
\keyword{none}

