% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HHrecs.R
\docType{data}
\name{HHrecs}
\alias{HHrecs}
\title{Haul-based information for ICES NS-IBTS surveys from 1965-2014}
\format{A data frame with 8234 rows and 29 variables:
\describe{
  \item{Survey}{price, in US dollars}
  \item{Quarter}{weight of the diamond, in carats}
  \item{Ship}{price, in US dollars}
  \item{Gear}{weight of the diamond, in carats}
  \item{SweepLngt}{price, in US dollars}
  \item{GearExp}{weight of the diamond, in carats}
  \item{DoorType}{price, in US dollars}
  \item{StNo}{weight of the diamond, in carats}
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
}}
\source{
\url{https://datras.ices.dk/WebServices/Webservices.aspx}
}
\usage{
HHrecs
}
\description{
A dataset containing haul-based information for ICES North Sea International
bottom trawl survey.
}
\keyword{datasets}

