% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HLrecs.R
\docType{data}
\name{HLrecs}
\alias{HLrecs}
\title{Length-based species information for ICES NS-IBTS surveys from 1965-2014}
\format{A data frame with 845052 rows and 25 variables:
\describe{
  \item{Survey}{price, in US dollars}
  \item{Quarter}{weight of the diamond, in carats}
  \item{Ship}{price, in US dollars}
  \item{Gear}{weight of the diamond, in carats}
  \item{SweepLngt}{price, in US dollars}
  \item{GearExp}{weight of the diamond, in carats}
  \item{DoorType}{price, in US dollars}
  \item{StNo}{weight of the diamond, in carats}
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
}}
\source{
\url{https://datras.ices.dk/WebServices/Webservices.aspx}
}
\usage{
HLrecs
}
\description{
A dataset containing length-based species information for ICES North Sea
international bottom trawl survey.
}
\keyword{datasets}

