% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cacheDATRAS.R
\name{loadCache}
\alias{loadCache}
\title{Load an updated cache}
\usage{
loadCache(path = NULL, surveyName)
}
\arguments{
\item{path}{location where cache is located}
}
\value{
loads the DATRAS survey object into the working space.
}
\description{
Load an updated cache
}
\examples{
\dontrun{
 catchDATRAS()
 loadDATRAS()
}
}
\seealso{
\code{\link{cacheDATRAS}} and \code{\link{getDATRAS}}
}
\keyword{cache}

