% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSpatial.R
\name{readSpatial}
\alias{readSpatial}
\title{readSpatial.R
Download and read data from the ICES Spatial Facility}
\usage{
readSpatial(dataset = c("ices_areas", "ices_rectangles", "ices_ecoregions",
  "helcom_subbasins", "ospar_regions"), ...)
}
\arguments{
\item{dataset, }{...}
}
\value{
none
}
\description{
readSpatial.R
Download and read data from the ICES Spatial Facility
}
\details{
Reads ICES Statistical areas, ICES Statistical Rectangles,ICES Ecoregions, HELCOM sub-basins, or OSPAR Regions as 'rgdal' SpatialPolygonsDataFrame objects.
  Note: You will be prompted to read and agree to the ICES Data Policy before the download will begin.
}
\examples{
\dontrun{
 readSpatial(dataset = "ices_areas")
}
}
\seealso{
none
}
\keyword{download,}
\keyword{spatial}

