% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cacheDATRAS.R
\name{updateCache}
\alias{updateCache}
\title{Update the cached copy of DATRAS data}
\usage{
updateCache(path = ".", surveyName, startYear, endYear, startQuarter,
  endQuarter, inParallel = FALSE)
}
\arguments{
\item{path}{where cache should be stored. (default to working directory)}
}
\value{
a date-database.Rdat file.
}
\description{
Update the cached copy of DATRAS data
}
\details{
the update is slow, avoiding straining the server or client.
  please allow this call to run overnight for a complete upgrade.
}
\examples{
\dontrun{
 updateCache(path = ".", surveyName = "NS-IBTS",
 startYear = 2005, endYear = 2010, startQuarter = 1, endQuarter = 4, inParallel = FALSE)
 loadCache()
}
}
\seealso{
\code{\link{loadDATRAS}}, and \code{\link{getDATRAS}}
}
\keyword{cache}

