% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDATRAS.R, R/getStomach.R
\name{getDATRAS}
\alias{getDATRAS}
\title{getDATRASfun.R
Extracts age data files from DATRAS}
\usage{
getDATRAS(record, survey, startyear, endyear, quarters, parallel = FALSE,
  cores = NULL)

getDATRAS(record, survey, startyear, endyear, quarters, parallel = FALSE,
  cores = NULL)
}
\arguments{
\item{record, }{survey, startyear, endyear, startquarter, endquarter, parallel = FALSE, cores = NULL}

\item{record, }{survey, startyear, endyear, startquarter, endquarter, parallel = FALSE, cores = NULL}
}
\value{
none

none
}
\description{
getDATRASfun.R
Extracts age data files from DATRAS

getStomach.R
The first year of the stomach ran in 1981 and covered a handful of species in the North
Sea. There was some follow up data collection performed in 1985-1986 and a decade
later in 1991 the 2nd year of the stomach ran. The data were collected as part of an ICES
initiative and the results and analysis were presented in two cooperative reports (CRR164 and CRR 219).
}
\details{
the update is slow, avoiding straining the server or client.
  please allow this call to run overnight for a complete upgrade.

the update is slow, avoiding straining the server or client.
  please allow this call to run overnight for a complete upgrade.
}
\examples{
\dontrun{
 getDATRAS(record = "HH",
           survey="NS-IBTS",
           startyear = 2010,
           endyear = 2011,
           quarters = 1,
           parallel = TRUE,
           cores = 4)
}
\dontrun{
 getDATRAS(record = "HH",
           survey="NS-IBTS",
           startyear = 2010,
           endyear = 2011,
           quarters = 1,
           parallel = TRUE,
           cores = 4)
}
}
\seealso{
none

none
}
\keyword{DATRAS,}
\keyword{age,}
\keyword{download,}
\keyword{length}
\keyword{survey,}

