% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDATRAS.R
\name{getDATRAS}
\alias{getDATRAS}
\title{getDATRAS}
\usage{
getDATRAS(surveyName, startYear, endYear, startQuarter, endQuarter,
  dataDir = "~/", inParallel = FALSE)
}
\arguments{
\item{surveyName,}{startYear, endYear, startQuarter, endQuarter, dataDir = "~/", inParallel}
}
\value{
none
}
\description{
Download files from DATRAS
}

