% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getHHfun.R
\name{getHHfun}
\alias{getHHfun}
\title{Haul function}
\usage{
getHHfun(survey, startyear, endyear, startquarter, endquarter,
  parallel = FALSE)
}
\arguments{
\item{survey,}{startyear, endyear, startquarter, endquarter, parallel}
}
\value{
none
}
\description{
More descriptive name
}

