% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getHLfun.R
\name{getHLfun}
\alias{getHLfun}
\title{Datras Function}
\usage{
getHLfun(survey, startyear, endyear, startquarter, endquarter,
  parallel = FALSE)
}
\arguments{
\item{x}{}
}
\value{
none
}
\description{
More descriptive name
}

