% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getStockSummary.R
\name{getSummaryTable}
\alias{getSummaryTable}
\title{getStockSummary.R
ICES standard graph data extraction}
\usage{
getSummaryTable(year = 2014)
}
\arguments{
\item{year}{}
}
\value{
none
}
\description{
getStockSummary.R
ICES standard graph data extraction
}
\details{
none
}
\examples{
\dontrun{
tt <- getSummaryTable()
}
}
\seealso{
none
}
\keyword{none}

