% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addLogo.R
\name{icesLogo}
\alias{icesLogo}
\title{addLogo.R
Adds ICES logo to a base R plot}
\usage{
icesLogo(logoType = c("acronym", "fullText"), x, y, size, alpha = 1)
}
\arguments{
\item{logoType}{= c("acronym", "fullText"), x, y, size, alpha = 1}
}
\value{
none
}
\description{
addLogo.R
Adds ICES logo to a base R plot
}
\details{
none
}
\examples{
\dontrun{
 needList <- c("png")
 new.packages <- needList[!(needList \%in\% installed.packages()[,"Package"])]
 if(length(new.packages)) install.packages(new.packages)
 library(png)

 width <- 172.4
 height <- 162.6
 rows <- 2
 cols <- 1
 png(filename = "~/TESTlogo.png",
     width = width,
     height = height,
     units = "mm",
     res = 600)

 par(mar = c(2.15, 2.25, 0.45, 0.25),
     mfrow = c(rows, cols),
     oma = c(0, 0, 1.25, 0),
     mgp = c(3, .35, 0),
     tck = -0.01,
     family = "Calibri")

 plot(1:2,type="n")
 icesLogo(logoType = "acronym", x = 0.1, y = 0.9, size = .25, alpha = 1)
 plot(1:2,type="n")
 icesLogo(logoType = "fullText", x = 0.5, y = 0.5, size = 1, alpha = .1)
 dev.off()


 png(filename = "~/rICESlogo.png",
 width = 25.4,
   height = 25.4,
   units = "mm",
   res = 600)
par(mar = c(0, 0, 0, 0),
   oma = c(0, 0, 0, 0))
orange <- "#F15D2A"
plot(x = 0.5, y = 0.64,type="p", pch = "R", cex = 4, col = orange, xlim = c(0,1), ylim = c(0,1),
    axes = F, bty = "n")
icesLogo(logoType = "acronym", x = 0.5, y = 0.5, size = .55, alpha = .9)
dev.off()
}
}
\seealso{
none
}
\keyword{plot}

